Kafka/Producer/src/requirements.txt
Random-Word
isbnlib
kafka-python